requirements.txt
click (6.6)
Flask (0.11.1)
itsdangerous (0.24)
Jinja2 (2.8)
MarkupSafe (0.23)
mysql-connector (2.1.4)
pip (1.5.6)
setuptools (3.6)
Werkzeug (0.11.11)
